==> Lessons/11_Lab_CraftingReports_Dashboards.Rmd <==
---
title: 'Lab 11: Crafting Reports and Dashboards in R'
author: Environmental Data Analytics | John Fay & Luana Lima | Developed by Kateri Salk
date: "Spring 2022"
output:
 pdf_document: default

---

## OBJECTIVES

1. Questions on M10/A09
2. Talk about final project (groups/prompt/rubric)
    Enter groups [here][https://docs.google.com/spreadsheets/d/1rI5c4dOzbmg3JRVJbkOEEIA7m5EtNTtslQbCw255cZ0/edit?usp=sharing] 
    
check for correct link on slack. 
    
3. Questions on M11
4. Exercise 1: Fill lesson file for M11
5. Exercise 2: Shiny app
